# deps: Upgrade github.com/gohugoio/hashstructure from 0.1.0 to 0.3.0